requirements.txt
fastapi
uvicorn[standard]
tensorflow
pillow
opencv-python-headless
numpy
mysql-connector-python
jinja2
